feat([sc-10626]): rename standard ICRC-33 to ICRC-49
