+++ docs/pages/blog/standardizing-icp-wallet-and-identity-interaction-overview.mdx
@@ -58,13 +58,13 @@ of applications, ensuring that user identities are protected and authenticated r
 
 [Technical draft](../identity-kit/icrc-32-sign-challenge)
 
-### ICRC-33 Call Canister
+### ICRC-49 Call Canister
 
 This standard allows applications to request the execution of calls to third-party canisters by the
 signer. Using the `icrc33_call_canister` method, applications can interact with a broader range of
 services and functionalities, enhancing the versatility and reach of ICP-based applications.
 
-[Technical draft](../identity-kit/icrc-33-call-canister)
+[Technical draft](../identity-kit/icrc-49-call-canister)
 
 ### ICRC-34 Get Delegation
