+++ docs/pages/identity-kit/icrc-49-call-canister.mdx
@@ -1,15 +1,15 @@
 ---
-title: "ICRC-33: Call Canister"
+title: "ICRC-49: Call Canister"
 date: "2023-12-12"
 authors:
   - name: "Dan Ostrovski"
   - name: "Philipp Litzenberger"
 ---
 
-import { Stack } from "../../components/stack.tsx"
+import { Stack } from "../../components/stack.jsx"
 import { Tabs } from "nextra/components"
 
-## ICRC-33 Call Canister
+## ICRC-49 Call Canister
 
 <Stack>
 [![Status Badge](https://img.shields.io/badge/STATUS-DRAFT-ffcc00.svg)](https://github.com/dfinity/wg-identity-authentication/blob/main/topics/icrc_33_call_canister.md)
